Переработал регистрацию на игу: 1. покрыл тестами функционал регистрации на игру 2. рефакторинг кода 3. добавил исключения ограничения доступа к орг.возможностям 4. доработка стилей 5. добавил meta viewport

+++ db/shvatka.sql
@@ -78,7 +78,7 @@ CREATE TABLE IF NOT EXISTS `sh_game` (
   `n` int(11) NOT NULL AUTO_INCREMENT,
   `uroven` int(11) NOT NULL DEFAULT '0',
   `n_podskazki` int(11) NOT NULL DEFAULT '0',
-  `p_time` int(11) NOT NULL DEFAULT '0',
+  `p_time` float NOT NULL DEFAULT '0',
   `text` text ,
   `keyw` varchar(100)  DEFAULT NULL,
   `b_keyw` varchar(100)  NOT NULL,
